Align precision, types, and constants with algorithm accuracy
- AstroReading: Unix event timestamps and MoonDistanceKilometres changed
  from double to long; long.MinValue is the sentinel for missing events
- PrometheusMetricsExporter: per-type formatting (2 dp angles, 4 dp
  fractions, integer for timestamps/distance); FormatInteger accepts long
- AstroCalculator: named constants replace all magic numbers
  (SecondsPerDay, JdUnixEpoch, DegreesPerHour, AzimuthSouthToNorthOffset,
  OneDayJd, MillisecondsPerSecond); helpers return long
- Snapshots regenerated to match new precision and types
- .editorconfig added with max_line_length=120
- CLAUDE.md: updated architecture description, added Markdown and coding
  conventions; README.md: added Precision section, prose wrapped at 120

🤖 AI Assisted with Claude Code

+++ AstroClockExporter.Core/Astro/AstroReading.cs
@@ -11,23 +11,23 @@ public sealed record AstroReading
     // Moon metadata
     public double MoonIlluminationFraction { get; init; }
     public double MoonPhaseAngleDegrees { get; init; }
-    public double MoonDistanceKilometres { get; init; }
+    public long MoonDistanceKilometres { get; init; }
 
-    // Sun events for today (Unix seconds, NaN if event doesn't occur)
-    public double SunriseUnix { get; init; }
-    public double SunsetUnix { get; init; }
-    public double SolarNoonUnix { get; init; }
-    public double CivilDawnUnix { get; init; }
-    public double CivilDuskUnix { get; init; }
-    public double NauticalDawnUnix { get; init; }
-    public double NauticalDuskUnix { get; init; }
-    public double AstronomicalDawnUnix { get; init; }
-    public double AstronomicalDuskUnix { get; init; }
+    // Sun events for today (Unix seconds, long.MinValue if event doesn't occur)
+    public long SunriseUnix { get; init; }
+    public long SunsetUnix { get; init; }
+    public long SolarNoonUnix { get; init; }
+    public long CivilDawnUnix { get; init; }
+    public long CivilDuskUnix { get; init; }
+    public long NauticalDawnUnix { get; init; }
+    public long NauticalDuskUnix { get; init; }
+    public long AstronomicalDawnUnix { get; init; }
+    public long AstronomicalDuskUnix { get; init; }
 
-    // Moon events for today (Unix seconds, NaN if event doesn't occur)
-    public double MoonriseUnix { get; init; }
-    public double MoonsetUnix { get; init; }
-    public double MoonTransitUnix { get; init; }
+    // Moon events for today (Unix seconds, long.MinValue if event doesn't occur)
+    public long MoonriseUnix { get; init; }
+    public long MoonsetUnix { get; init; }
+    public long MoonTransitUnix { get; init; }
 
     // Self-monitoring
     public double CalculationSeconds { get; init; }
